rename: Genres Components 중 ContentsBox 상속받는 Component 이동
