+++ src/screens/MainBox/Genres/Components/ContentsBox/RecommendEvent.js
@@ -1,4 +1,4 @@
-import { createExtendsRelation } from "../../../../utils.js";
+import { createExtendsRelation } from "../../../../../utils.js";
 import ContentsBox from "../ContentsBox.js";
 
 function RecommendEvent(infoObject) {
