+++ src/screens/MainBox/Genres/Webtoon/genre/HomeGenre.js
@@ -1,17 +1,17 @@
 import { createExtendsRelation } from "../../../../../utils.js";
 import { getComponentsTemplate } from "../../../../../modules/serviceUtils.js";
-import ContentsBox from "../../../Genres/ContentsBox.js";
 import BigCardList from "../../Components/BigCardList.js";
-import DateTop from "../../Components/DateTop.js";
+import DateTop from "../../Components/ContentsBox/DateTop.js";
 import DaysTop from "../../Components/DaysTop.js";
-import GenreTop from "../../Components/GenreTop.js";
+import GenreTop from "../../Components/ContentsBox/GenreTop.js";
 import MainBanner from "../../Components/MainBanner.js";
 import NavDetail from "../../Components/NavDetail.js";
-import RecommendEvent from "../../Components/RecommendEvent.js";
 import SubBanner from "../../Components/SubBanner.js";
 import Component from "../../../../Component.js";
 import DaysList from "../../Components/DaysList.js";
 import FullButton from "../../Components/FullButton.js";
+import ContentsBox from "../../Components/ContentsBox.js";
+import RecommendEvent from "../../Components/ContentsBox/RecommendEvent.js";
 
 function HomeGenre(target) {
   Component.call(this, target);
