+++ src/screens/MainBox/Genres/components.js
@@ -1,12 +1,12 @@
 import BigCardList from "./Components/BigCardList.js";
-import DateTop from "./Components/DateTop.js";
 import DaysTop from "./Components/DaysTop.js";
 import FullButton from "./Components/FullButton.js";
-import GenreTop from "./Components/GenreTop.js";
 import MainBanner from "./Components/MainBanner.js";
 import NavDetail from "./Components/NavDetail";
-import RecommendEvent from "./Components/RecommendEvent.js";
 import SubBanner from "./Components/SubBanner.js";
+import DateTop from "./Components/ContentsBox/DateTop.js";
+import GenreTop from "./Components/ContentsBox/GenreTop.js";
+import RecommendEvent from "./Components/ContentsBox/RecommendEvent.js";
 
 const components = {
   mainBanner: MainBanner,
